backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-huggingface
sentence-transformers
faiss-cpu
pymupdf
docx2txt
tiktoken